requirements.txt
streamlit
chromadb
langchain
langchain-community
langchain-core
langchain-groq
langchain-huggingface
sentence-transformers
fastapi
uvicorn
huggingface_hub
python-dotenv
